+++ network-api/networkapi/wagtailpages/templates/wagtailpages/tags/cta/petition.html
@@ -28,12 +28,7 @@
         data-share-link="{{ cta.share_link }}"
         data-share-text="{{ cta.share_link_text }}"
         data-thank-you="{{ cta.thank_you }}"
-
-        data-modal-name="{{ modal.name }}"
-        data-modal-header="{{ modal.header }}"
-        data-modal-body="{{ modal.body }}"
-        data-modal-donate-text="{{ modal.donate_text }}"
-        data-modal-dismiss-text="{{ modal.dismiss_text }}"
+        data-modals="{{ modals_json }}"
 	>loading CTA</div>
 
     <script src="https://c.shpg.org/352/sp.js"></script>
